+++ packages/mecha/src/index.ts
@@ -1,0 +1,6 @@
+export type {
+  AgentConfigResolverContext,
+  AgentConfigSnapshot,
+  ResolvedAgentConfig,
+} from './resolveAgentConfig';
+export { resolveAgentConfig } from './resolveAgentConfig';
